service-workers/service-worker/clients-matchall-client-types.https-expected.txt

This is a testharness.js-based test.
PASS Verify matchAll() with window client type 
FAIL Verify matchAll() with {window, sharedworker, worker} client types assert_equals: expecting ,,https://web-platform.test:8444/service-workers/service-worker/resources/clients-matchall-client-types-dedicated-worker.js,worker,none: result count expected 1 but got 0
Harness: the test ran to completion.
